block_buffer::sealed

Trait Sealed

pub trait Sealed {
    type Pos: Default + Clone + Zeroize;
    type Overhead: ArraySize;

    const NAME: &'static str;

    // Required methods
    fn get_pos<N: ArraySize>(
        buf: &MaybeUninit<Array<u8, N>>,
        pos: &Self::Pos,
    ) -> usize;
    fn set_pos<N: ArraySize>(
        buf: &mut MaybeUninit<Array<u8, N>>,
        pos: &mut Self::Pos,
        val: usize,
    );
    fn invariant(pos: usize, block_size: usize) -> bool;
    fn split_blocks<N: ArraySize>(data: &[u8]) -> (&[Array<u8, N>], &[u8]);
}

Sealed trait for buffer kinds.

Required Associated Constants

const NAME: &'static str

Required Associated Types

type Pos: Default + Clone + Zeroize

type Overhead: ArraySize

Required Methods

fn get_pos<N: ArraySize>( buf: &MaybeUninit<Array<u8, N>>, pos: &Self::Pos, ) -> usize

fn set_pos<N: ArraySize>( buf: &mut MaybeUninit<Array<u8, N>>, pos: &mut Self::Pos, val: usize, )

fn invariant(pos: usize, block_size: usize) -> bool

Invariant guaranteed by a buffer kind, i.e. with correct buffer code this function always returns true.

fn split_blocks<N: ArraySize>(data: &[u8]) -> (&[Array<u8, N>], &[u8])

Split input data into slice of blocks and tail.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.