Struct block_buffer::BlockBuffer

pub struct BlockBuffer<BlockSize: ArrayLength<u8>> { /* fields omitted */ }

Buffer for block processing of data

Methods

impl<BlockSize: ArrayLength<u8>> BlockBuffer<BlockSize>

pub fn input<F>(&mut self, input: &[u8], f: F) where
F: FnMut(&GenericArray<u8, BlockSize>),

Process data in input in blocks of size BlockSize using function f.

pub fn input_lazy<F>(&mut self, input: &[u8], f: F) where
F: FnMut(&GenericArray<u8, BlockSize>),

Variant that doesn't flush the buffer until there's additional data to be processed. Suitable for tweakable block ciphers like Threefish that need to know whether a block is the last data block before processing it.

pub fn len64_padding<B, F>(&mut self, data_len: u64, f: F) where
B: ByteOrder,
F: FnMut(&GenericArray<u8, BlockSize>),

Pad message with 0x80, zeros and 64-bit message length in a byte order specified by B

pub fn len128_padding_be<F>(&mut self, hi: u64, lo: u64, f: F) where
F: FnMut(&GenericArray<u8, BlockSize>),

Pad message with 0x80, zeros and 128-bit message length in the big-endian byte order

pub fn pad_with<P: Padding>(
&mut self
) -> Result<&mut GenericArray<u8, BlockSize>, PadError>

Pad message with a given padding P

Returns PadError if internall buffer is full, which can only happen if input_lazy was used.

pub fn size(&self) -> usize

Return size of the internall buffer in bytes

pub fn position(&self) -> usize

Return current cursor position

pub fn remaining(&self) -> usize

Return number of remaining bytes in the internall buffer

pub fn reset(&mut self)

Reset buffer by setting cursor position to zero