block_buffer

Struct ReadBuffer

pub struct ReadBuffer<BS: BlockSizes> { /* private fields */ }

Buffer for reading block-generated data.

Implementations

impl<BS: BlockSizes> ReadBuffer<BS>

pub fn get_pos(&self) -> usize

Return current cursor position, i.e. how many bytes were read from the buffer.

pub fn size(&self) -> usize

Return size of the internal buffer in bytes.

pub fn remaining(&self) -> usize

Return number of remaining bytes in the internal buffer.

pub fn read_cached(&mut self, len: usize) -> &[u8]

Read up to len bytes of remaining data in the buffer.

Returns slice with length of ret_len = min(len, buffer.remaining()) bytes and sets the cursor position to buffer.get_pos() + ret_len.

pub fn write_block( &mut self, read_len: usize, gen_block: impl FnOnce(&mut Array<u8, BS>), read_fn: impl FnOnce(&[u8]), )

Write new block and consume read_len bytes from it.

If read_len is equal to zero, immediately returns without calling the closures. Otherwise, the method calls gen_block to fill the internal buffer, passes to read_fn slice with first read_len bytes of the block, and sets the cursor position to read_len.

Panics

If read_len is bigger than block size.

pub fn reset(&mut self)

Reset buffer into exhausted state.

pub fn read( &mut self, buf: &mut [u8], gen_block: impl FnMut(&mut Array<u8, BS>), )

Write remaining data inside buffer into buf, fill remaining space in buf with blocks generated by gen_block, and save leftover data from the last generated block into the buffer for future use.

pub fn serialize(&self) -> Array<u8, BS>

Serialize buffer into a byte array.

pub fn deserialize(buffer: &Array<u8, BS>) -> Result<Self, Error>

Deserialize buffer from a byte array.

Errors

If the first byte is 0.

If the first byte is bigger than BS.