Trait block_buffer::Unpadding

pub trait Unpadding {
    fn unpad(data: &[u8]) -> Result<&[u8], UnpadError>;
}

Trait for extracting oringinal message from padded medium

Required Methods

fn unpad(data: &[u8]) -> Result<&[u8], UnpadError>

Unpad given data by truncating it according to the used padding. In case of the malformed padding will return UnpadError