Type Alias block_buffer::LazyBuffer

pub type LazyBuffer<B> = BlockBuffer<B, Lazy>;

Lazy block buffer.

Aliased Type

struct LazyBuffer<B> { /* private fields */ }

Implementations

impl<BlockSize, Kind> BlockBuffer<BlockSize, Kind>where BlockSize: ArrayLength<u8> + IsLess<U256>, Le<BlockSize, U256>: NonZero, Kind: BufferKind,

pub fn new(buf: &[u8]) -> Self

Create new buffer from slice.

Panics

If slice length is not valid for used buffer kind.

pub fn try_new(buf: &[u8]) -> Result<Self, Error>

Create new buffer from slice.

Returns an error if slice length is not valid for used buffer kind.

pub fn digest_blocks( &mut self, input: &[u8], compress: impl FnMut(&[Block<BlockSize>]) )

Digest data in input in blocks of size BlockSize using the compress function, which accepts slice of blocks.

pub fn reset(&mut self)

Reset buffer by setting cursor position to zero.

pub fn pad_with_zeros(&mut self) -> &mut Block<BlockSize>

Pad remaining data with zeros and return resulting block.

pub fn get_pos(&self) -> usize

Return current cursor position.

pub fn get_data(&self) -> &[u8]

Return slice of data stored inside the buffer.

pub fn set(&mut self, buf: Block<BlockSize>, pos: usize)

Set buffer content and cursor position.

Panics

If pos is bigger or equal to block size.

pub fn size(&self) -> usize

Return size of the internal buffer in bytes.

pub fn remaining(&self) -> usize

Return number of remaining bytes in the internal buffer.