Crate block_buffer

Fixed size buffer for block processing of data.

Re-exports

pub use generic_array;

Structs

BlockBuffer

Buffer for block processing of data.

Eager

Eager block buffer kind, which guarantees that buffer position always lies in the range of 0..BlockSize.

Lazy

Lazy block buffer kind, which guarantees that buffer position always lies in the range of 0..=BlockSize.

Traits

BufferKind

Trait for buffer kinds.

Type Definitions

Block

Block on which BlockBuffer operates.

EagerBuffer

Eager block buffer.

LazyBuffer

Lazy block buffer.