Trait block_buffer::sealed::Sealed

pub trait Sealed {
    // Required methods
    fn invariant(pos: usize, block_size: usize) -> bool;
    fn split_blocks<N: ArrayLength<u8>>(data: &[u8]) -> (&[Block<N>], &[u8]);
}

Sealed trait for buffer kinds.

Required Methods

fn invariant(pos: usize, block_size: usize) -> bool

Invariant guaranteed by a buffer kind, i.e. with correct buffer code this function always returns true.

fn split_blocks<N: ArrayLength<u8>>(data: &[u8]) -> (&[Block<N>], &[u8])

Split input data into slice of blocks and tail.